// frontend/src/api/generated/models/itemRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Trippy API
 * OpenAPI spec version: v1
 */
import type { PlaceRead } from './placeRead';

export interface ItemRead {
  id: string;
  section_id: string;
  /** @minLength 1 */
  place_id: string;
  /** @minLength 1 */
  description: string;
  sequence: number;
  /** @nullable */
  start_time?: string | null;
  place?: PlaceRead;
}